+++ packages/riverpod/lib/src/future_provider/auto_dispose.dart
@@ -133,6 +133,8 @@ class AutoDisposeFutureProviderFamily<State, Arg>
       name: name,
       from: this,
       argument: argument,
+      cacheTime: cacheTime,
+      disposeDelay: disposeDelay,
     );
   }