+++ packages/riverpod/lib/src/provider/auto_dispose.dart
@@ -94,6 +94,8 @@ class AutoDisposeProviderFamily<State, Arg>
       name: name,
       from: this,
       argument: argument,
+      cacheTime: cacheTime,
+      disposeDelay: disposeDelay,
     );
   }
 }