+++ packages/riverpod/lib/src/state_notifier_provider/auto_dispose.dart
@@ -233,6 +233,8 @@ class AutoDisposeStateNotifierProviderFamily<
       name: name,
       from: this,
       argument: argument,
+      cacheTime: cacheTime,
+      disposeDelay: disposeDelay,
     );
   }