+++ packages/riverpod/lib/src/state_provider/auto_dispose.dart
@@ -188,6 +188,8 @@ class AutoDisposeStateProviderFamily<State, Arg>
       name: name,
       from: this,
       argument: argument,
+      cacheTime: cacheTime,
+      disposeDelay: disposeDelay,
     );
   }