+++ packages/riverpod/lib/src/stream_provider/auto_dispose.dart
@@ -145,6 +145,8 @@ class AutoDisposeStreamProviderFamily<State, Arg>
       name: name,
       from: this,
       argument: argument,
+      cacheTime: cacheTime,
+      disposeDelay: disposeDelay,
     );
   }