+++ packages/riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `AsyncValue.valueOrNull` to obtain the value while ignoring potential errors.
 - Fixed an issue where `AsyncValue.value` did not throw if there is an error.
+- Fix families not applying cacheTime/disposeDelay
 
 # 2.0.0-dev.5